+++ client/src/App.jsx
@@ -1,8 +1,8 @@
 import { useState } from 'react'
-import reactLogo from './assets/react.svg'
+//import reactLogo from './assets/react.svg'
 import viteLogo from '/vite.svg'
 import './App.css'
-import Homepage from './Homepage'
+import Homepage from './pages/Homepage'
 import { BrowserRouter, Routes, Route } from "react-router-dom";
 
 function App() {
